feat: add RPC and declare request

+++ create-summer/index.js
@@ -31,9 +31,18 @@ var copyRecursiveSync = function (src, dest) {
   if (!projectName) {
     return
   }
+
+  if (fs.existsSync(projectName)) {
+    console.log(projectName + ' exists, exit')
+    return
+  }
+
   copyRecursiveSync(path.join(__dirname, 'template'), projectName)
 
   fs.writeFileSync(projectName + '/.gitignore', ['.DS_Store', 'node_modules', 'build', 'compile'].join('\n'))
+  const packageJson = JSON.parse(fs.readFileSync(projectName + '/package.json', { encoding: 'utf8' }))
+  packageJson.name = projectName
+  fs.writeFileSync(projectName + '/package.json', JSON.stringify(packageJson, null, 4))
 
   console.log('Project Created! Now run:')
   console.log('')
